refactor(connector): extract provider types and handlers into separate files
