+++ turbo/apps/web/src/lib/connector/providers/github-handler.ts
@@ -1,0 +1,24 @@
+import { type ProviderHandler } from "../provider-types";
+import {
+  buildGitHubAuthorizationUrl,
+  exchangeGitHubCode,
+  fetchGitHubUserInfo,
+  getGitHubSecretName,
+} from "./github";
+
+export const githubHandler: ProviderHandler = {
+  buildAuthUrl: buildGitHubAuthorizationUrl,
+  async exchangeCode(clientId, clientSecret, code, redirectUri) {
+    const { accessToken, scopes } = await exchangeGitHubCode(
+      clientId,
+      clientSecret,
+      code,
+      redirectUri,
+    );
+    const userInfo = await fetchGitHubUserInfo(accessToken);
+    return { accessToken, scopes, userInfo };
+  },
+  getClientId: (e) => e.GH_OAUTH_CLIENT_ID,
+  getClientSecret: (e) => e.GH_OAUTH_CLIENT_SECRET,
+  getSecretName: getGitHubSecretName,
+};
